+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 * Keep unhandled WAVE subchunks.
 * Store ``ID3v2`` in ``FLAC`` in ``_id3`` attribute.
 * Emit warning when ignoring ID3v2 frames with bad values.
+* Ignore ID3v2 frames with no value.
 
 ### Fixed